requirements.txt
flask
numpy
opencv-python-headless
pytesseract
pillow
requests